changed filerev_replace task
Updated the filerev_replace task to a single dist entry against all html, css and js files. Ran nodeunit and am close to passing all tests. The relative 'images/ajax-loader.gif' path in 3 of the expected is different as my code replaces that path. However, my testing of the previous code shows much more wrong in a diff between the output to tmp and expected.